+++ src/main/java/com/backend/girnartour/RequestDTOs/SalesHeaderRequest.java
@@ -6,6 +6,8 @@ import com.backend.girnartour.models.User;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
+
+import java.math.BigDecimal;
 import java.util.List;
 
 @Data
@@ -22,13 +24,13 @@ public class SalesHeaderRequest {
 
     public String message;
 
-    public Double discount;
+    public BigDecimal discount;
 
-    public Double vatAmt;
+    public BigDecimal vatAmt;
 
-    public Double invoiceAmt;
+    public BigDecimal invoiceAmt;
 
-    public Double totalInvoiceAmt;
+    public BigDecimal totalInvoiceAmt;
 
     public List<SalesDetail> salesDetailList;
